rename firewall module
